Move test in with other archive reference tests
